refactor: Extract type utility functions to typeUtils.ts
Break circular dependency: file.ts → @types (index.ts) → knowledge.ts → file.ts
by moving objectKeys, objectEntries, objectEntriesStrict, objectValues, strip,
AtLeast, RequireSome, NotUndefined, NotNull to a dedicated typeUtils.ts module.
