+++ src/renderer/src/types/file.ts
@@ -1,8 +1,9 @@
 import type OpenAI from '@cherrystudio/openai'
 import type { File } from '@google/genai'
 import type { FileSchema } from '@mistralai/mistralai/models/components'
-import { objectValues } from '@types'
 import * as z from 'zod'
+
+import { objectValues } from './typeUtils'
 
 export type RemoteFile =
   | {
